moved App.js to Sign-Up file.
