+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/index.js":"1","/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/App.js":"2","/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/reportWebVitals.js":"3","/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/components/SignUp.js":"4","/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/validation/SUformSchema.js":"5","/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/theme/index.js":"6"},{"size":727,"mtime":1612215228725,"results":"7","hashOfConfig":"8"},{"size":1849,"mtime":1612217686754,"results":"9","hashOfConfig":"8"},{"size":362,"mtime":1612201805416,"results":"10","hashOfConfig":"8"},{"size":3764,"mtime":1612218499676,"results":"11","hashOfConfig":"8"},{"size":887,"mtime":1612215233925,"results":"12","hashOfConfig":"8"},{"size":0,"mtime":1612215373901,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1vzrl0e",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/index.js",[],"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/App.js",[],"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/reportWebVitals.js",[],"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/components/SignUp.js",[],"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/validation/SUformSchema.js",[],"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/theme/index.js",[]]
+[{"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/index.js":"1","/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/App.js":"2","/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/reportWebVitals.js":"3","/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/components/SignUp.js":"4","/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/validation/SUformSchema.js":"5","/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/theme/index.js":"6"},{"size":727,"mtime":1612219064754,"results":"7","hashOfConfig":"8"},{"size":249,"mtime":1612219056630,"results":"9","hashOfConfig":"8"},{"size":362,"mtime":1612201805416,"results":"10","hashOfConfig":"8"},{"size":5222,"mtime":1612219103094,"results":"11","hashOfConfig":"8"},{"size":887,"mtime":1612215233925,"results":"12","hashOfConfig":"8"},{"size":0,"mtime":1612219063647,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1vzrl0e",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/index.js",[],"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/App.js",[],"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/reportWebVitals.js",[],"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/components/SignUp.js",[],"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/validation/SUformSchema.js",[],["27","28"],"/Users/mzwissler/Documents/Lambda/Unit2BuildWeek.Web39/Frontend/client/src/theme/index.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
